requirements.txt
anyio==4.12.1
asgiref==3.11.1
certifi==2026.1.4
Django==5.2.11
django-allauth==65.14.3
django-rest-framework==0.1.0
djangorestframework==3.16.1
exceptiongroup==1.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.11
pillow==12.1.1
python-telegram-bot==20.7
sniffio==1.3.1
sqlparse==0.5.5
typing_extensions==4.15.0
